Fix bufferStorage
The current packed struct is wrong, it was implemented in the order
which the flags are described in the OpenGL Spec. This is not how the
bits are laid out in memory though, in fact if you were to write true to
map_write with this implementation it would through a GL_INVALID_VALUE
error since the function does not allow setting any bits besides the
defined ones. This commit fixes the bitfield to have the correct order
and padding aswell as providing their offsets as a comment and a small
description of what each flag does.